Notification: Fix no icon shown for data limit notification
When USB connects and data warning limit are enabled, the icon
for data limit is not shown in status bar after data limit reaches.

The reason is that when data limit notification reaches, the total
number of notifications for package android becomes 4, then will
auto-group, the group summary will re-use the firstly enqueued
notification's channel and importance, USB Charging notification is
the firstly enqueued one whose channel has IMPORTANT_MIN importance,
SystemUI will not show the icon for this unimportant notification,
it's considered as ambient notification in SystemUI.

For notifications in package android, they have different icons and
importance, so will not auto-group them.

Change-Id: Ia0792becea440c48fcb9a2adbdbdca674832e4b3
Signed-off-by: DennySPb <dennyspb@gmail.com>
Signed-off-by: niteshkumar2000 <nitesh156200@gmail.com>